requirements.txt
playwright
requests
python-dotenv

pywhatkit
twilio
